+++ src/align/reject_outliers/mod.rs
@@ -1,9 +1,7 @@
 mod keep_all;
-mod reject_3_sigmal_dist;
-mod reject_outliers_plane_dist;
+mod reject_n_sigma_dist;
 mod reject_overlapping_ratios;
 
 pub use keep_all::*;
-pub use reject_3_sigmal_dist::*;
-pub use reject_outliers_plane_dist::*;
+pub use reject_n_sigma_dist::*;
 pub use reject_overlapping_ratios::*;