+++ src/align/transform_estimation/mod.rs
@@ -1,5 +1,6 @@
 pub mod affine_transformation;
 pub mod point_to_plane_lls;
+pub mod point_to_plane_lls_weighted;
 pub mod scale;
 pub mod svd;
 pub mod transform_modifier;