+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2025-07-11
+
+- Renamed `reject_3_sigma_dist` to `reject_n_sigma_dist` to allow more flexibility in outlier rejection.
+- Added field `weight` to `PointCloudPoint` to allow weighted point cloud processing.
+- Added `point_to_plane_lls_weighted` which is the only transform estimator that respects the weights of the points.
+- Removed `reject_outliers_plane_dist`
+- Added `Plane::fit_to_point_cloud_wo_outliers`
+
 ## [0.4.0] - 2025-07-05
 
 - Added trait `ToPointCloud` to make ICP easy to use with point clouds loaded from various sources.